refactor(perl-lsp-rs): extract runtime support types module (#0000)
